add storage interface
