+++ go/worker/activities/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//go/worker/activities/ray:go_default_library",
-        "//go/worker/activities/s3:go_default_library",
+        "//go/worker/activities/storage:go_default_library",
         "@org_uber_go_fx//:go_default_library",
     ],
 )
