+++ go/worker/activities/module.go
@@ -2,11 +2,11 @@ package activities
 
 import (
 	"github.com/michelangelo-ai/michelangelo/go/worker/activities/ray"
-	"github.com/michelangelo-ai/michelangelo/go/worker/activities/s3"
+	"github.com/michelangelo-ai/michelangelo/go/worker/activities/storage"
 	"go.uber.org/fx"
 )
 
 var Module = fx.Options(
 	ray.Module,
-	s3.Module,
+	storage.Module,
 )
