+++ go/worker/activities/storage/BUILD.bazel
@@ -3,21 +3,21 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
-        "config.go",
+        "activities.go",
+        "interface.go",
         "module.go",
-        "s3_activities.go",
     ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/activities/s3",
+    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/activities/storage",
     visibility = ["//visibility:public"],
     deps = [
+        "//go/worker/activities/storage/minio:go_default_library",
+        "//proto/api/v2:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//ext:go_default_library",
         "@com_github_json_iterator_go//:go_default_library",
         "@com_github_minio_minio_go_v7//:go_default_library",
-        "@com_github_minio_minio_go_v7//pkg/credentials:go_default_library",
         "@org_uber_go_cadence//:go_default_library",
         "@org_uber_go_cadence//activity:go_default_library",
         "@org_uber_go_cadence//worker:go_default_library",
-        "@org_uber_go_config//:go_default_library",
         "@org_uber_go_fx//:go_default_library",
         "@org_uber_go_yarpc//yarpcerrors:go_default_library",
         "@org_uber_go_zap//:go_default_library",
