+++ go/worker/activities/storage/interface.go
@@ -2,11 +2,10 @@ package storage
 
 import (
 	"context"
-	"go.uber.org/cadence"
 )
 
 type Storage interface {
-	Read(ctx context.Context, req interface{}) (any, *cadence.CustomError)
+	Read(ctx context.Context, path string) (any, error)
 
 	Protocol() string
 }
