--- go/worker/plugins/s3/BUILD.bazel
@@ -1,18 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["plugin.go"],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/plugins/s3",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//go/worker/activities/s3:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//cadstar:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//ext:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//star:go_default_library",
-        "@com_github_json_iterator_go//:go_default_library",
-        "@net_starlark_go//starlark:go_default_library",
-        "@org_uber_go_cadence//worker:go_default_library",
-        "@org_uber_go_cadence//workflow:go_default_library",
-    ],
-)
