+++ go/worker/plugins/storage/BUILD.bazel
@@ -2,14 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "plugin.go",
-        "starlark_module.go",
-    ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/plugins/s3",
+    srcs = ["plugin.go"],
+    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/plugins/storage",
     visibility = ["//visibility:public"],
     deps = [
-        "//go/worker/activities/s3:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//cadstar:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//ext:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//star:go_default_library",
