+++ go/worker/workflows/BUILD.bazel
@@ -7,10 +7,6 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//go/worker/workflows/ray:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//cadstar:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//plugin:go_default_library",
-        "@org_uber_go_cadence//encoded:go_default_library",
         "@org_uber_go_fx//:go_default_library",
-        "@org_uber_go_zap//:go_default_library",
     ],
 )
